增加代码

+++ support-permission/src/main/java/com/hanshow/support/permission/dao/AuthoritiesRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.hanshow.support.permission.model.Authorities;
 
 @Repository
-public interface AuthoritiesRepository extends BaseRepository<Authorities, Long> {
+public interface AuthoritiesRepository extends BaseRepository<Authorities, Integer> {
 
 }
